Move ramp controllers to own file
